+++ neuron.h
@@ -6,34 +6,33 @@
 
 class Neuron {
 
-
 public:
 
 	Neuron(int _nInputs);
 	~Neuron();
 	void calcOutput();
 	void doLearning();
-	void initWeights(float _max);
-	inline float getOutput() { return output; };
-	inline float getSum() { return sum; };
-	inline float getWeight(int _index) { return weights[_index]; };
-	inline void setWeight(int _index, float _weight) { weights[_index]=_weight; };
-	inline void setError(float _error) { error=_error; };
-	inline float getError() { return error; };
-	inline void setInput(int _index, float _value) { inputs[_index] = _value; };
-	inline float getInput(int _index) { return inputs[_index]; };
-	inline void setBias(float _bias) { bias=_bias; };
-	inline void setLearningRate(float _learningrate) { learningRate = _learningrate; };
+	void initWeights(double _max);
+	inline double getOutput() { return output; };
+	inline double getSum() { return sum; };
+	inline double getWeight(int _index) { return weights[_index]; };
+	inline void setWeight(int _index, double _weight) { weights[_index]=_weight; };
+	inline void setError(double _error) { error=_error; };
+	inline double getError() { return error; };
+	inline void setInput(int _index, double _value) { inputs[_index] = _value; };
+	inline double getInput(int _index) { return inputs[_index]; };
+	inline void setBias(double _bias) { bias=_bias; };
+	inline void setLearningRate(double _learningrate) { learningRate = _learningrate; };
 
 private:
 	int nInputs;
-	float* weights;
-	float* inputs;
-	float output;
-	float sum;
-	float bias;
-	float error;
-	float learningRate;
+	double* weights;
+	double* inputs;
+	double output;
+	double sum;
+	double bias;
+	double error;
+	double learningRate;
 };
 
 #endif
